Remove intermediate sanity directory in the cms directory.
