+++ frontend/app/components/LiveVisualEditing.tsx
@@ -1,7 +1,7 @@
 import { VisualEditing } from "@sanity/visual-editing/remix";
 
-import { client } from "../../cms/sanity/clientConfig";
-import { useLiveMode } from "../../cms/sanity/loader";
+import { client } from "../../cms/clientConfig";
+import { useLiveMode } from "../../cms/loader";
 
 export default function LiveVisualEditing() {
   useLiveMode({ client });
