+++ frontend/app/routes/resource.preview.ts
@@ -5,7 +5,7 @@ import {
   redirect,
 } from "@remix-run/node";
 import { validatePreviewUrl } from "@sanity/preview-url-secret";
-import { client } from "../../cms/sanity/clientConfig";
+import { client } from "../../cms/clientConfig";
 
 import { commitSession, destroySession, getSession } from "../sessions";
