+++ frontend/cms/clientConfig.server.ts
