+++ frontend/cms/customTypes.ts
@@ -1,4 +1,4 @@
-import {} from "../../sanity.types";
+import {} from "../sanity.types";
 
 /***
  * This file is a custom type definition for the Sanity schema.
