+++ frontend/cms/schemaTypes/frontpage.ts
