+++ frontend/cms/schemaTypes/infopage.ts
