+++ frontend/cms/schemaTypes/objects/customImage.ts
