+++ frontend/cms/schemaTypes/objects/imageMask.ts
